loginFinition
finition du login et création des page pour gérer les utilisateurs avec le superAdmin

+++ view/page/teacher/list.php
@@ -3,11 +3,16 @@
 	<h2>Liste des enseignants</h2>
 	<div>
 		<?php
-			var_dump($_SESSION);
+			//var_dump($_SESSION);
 
-			if (array_key_exists("loged_in", $_SESSION) && $_SESSION["loged_in"])
+			if (array_key_exists("loged_in", $_SESSION) && $_SESSION["loged_in"]) // TODO : à vérifier s'il y a besoin de droit pour ajouter un prof
 			{
 				echo '<a href="index.php?controller=teacher&action=addTeacher"><button class="pull-right">ajouter un professeur</button></a>';
+
+				if (array_key_exists("userPermissionsNumber", $_SESSION) && $_SESSION["userPermissionsNumber"] >= 100)
+				{
+					echo '<a href="index.php?controller=user&action=manageUsers"><button class="pull-right">manage Users</button></a>';
+				}
 			}
 			else
 			{
@@ -36,8 +41,12 @@
 				if ( array_key_exists("loged_in", $_SESSION) && $_SESSION["loged_in"])
 				{
 					echo '<a href="index.php?controller=teacher&action=detail&id=' . htmlspecialchars($teacher['idTeacher']) . '"><img src="resources/image/iconLoupe.png" alt="image de loupe pour obtenir des informations supplémentaire sur l\'enseignant"></a>';
-					echo '<a onclick="return confirm(\'Voulez-vous vraiment supprimer cette entrée ?\')" href="deleteTeacher.php?id=' . htmlspecialchars($teacher['idTeacher']) . '"><img src="resources/image/iconTrash.png" alt="image de poubelle pour supprimer l\'enseignant de la base de donnée"></a>';
-					echo '<a href="index.php?controller=teacher&action=editTeacher&id=' . htmlspecialchars($teacher['idTeacher']) . '"><img src="resources/image/iconPencil.png" alt="image de loupe pour obtenir des informations supplémentaire sur l\'enseignant"></a>';
+					
+					if (array_key_exists("userPermissionsNumber", $_SESSION) && $_SESSION["userPermissionsNumber"] >= 75)
+					{
+						echo '<a onclick="return confirm(\'Voulez-vous vraiment supprimer cette entrée ?\')" href="deleteTeacher.php?id=' . htmlspecialchars($teacher['idTeacher']) . '"><img src="resources/image/iconTrash.png" alt="image de poubelle pour supprimer l\'enseignant de la base de donnée"></a>';
+						echo '<a href="index.php?controller=teacher&action=editTeacher&id=' . htmlspecialchars($teacher['idTeacher']) . '"><img src="resources/image/iconPencil.png" alt="image de loupe pour obtenir des informations supplémentaire sur l\'enseignant"></a>';
+					}
 				}
 				
 				echo '</td></tr>';
